update: variadic (002)
